requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.2
django-cors-headers==4.0.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-shortuuidfield==0.1.3
django-unixtimestampfield==0.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elastic-transport==8.4.0
elasticsearch==7.7.0
idna==3.4
Jinja2==3.1.2
json2==0.4.0
MarkupSafe==2.1.2
mysql-connector-python==8.0.33
Pillow==10.0.0
playsound==1.3.0
protobuf==3.20.3
psycopg2==2.9.6
pycparser==2.21
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
rest-framework-simplejwt==0.0.2
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
timedelta==2020.12.3
tzdata==2023.3
urllib3==1.26.16
